renaming hooks , updated project component schema, added project display toggle
