+++ studio/.env.production
@@ -1,0 +1,4 @@
+SANITY_STUDIO_PROJECT_ID=57e12m54
+SANITY_STUDIO_DATASET=badger3000
+SANITY_STUDIO_API_ID=82748cae-cf44-4f28-bceb-22e83e5c6e3f
+SANITY_STUDIO_API_ID_CMS=47abe777-1a82-44c6-8fb6-5a08aef31fb9
