+++ web/src/components/Hero/Hero.jsx
@@ -4,7 +4,7 @@ import { PortableText } from '@portabletext/react'
 export default function Hero(props) {
   return (
     <section>
-      <header className="">{props.heading}</header>
+      <header>{props.heading}</header>
       <PortableText value={props.hero_text} />
     </section>
   )
