+++ web/src/components/ProjectComponent/ProjectComponent.jsx
@@ -1,9 +1,17 @@
 import React from 'react'
-
+import Gallery from '../Gallery'
 const ProjectComponent = (props) => {
+  if (props.display_projects === true) {
+    return (
+      <section>
+        <h2>{props.heading}</h2>
+        <Gallery />
+      </section>
+    )
+  }
   return (
     <section>
-      <p>This is the project component</p>
+      <h2>{props.heading}</h2>
     </section>
   )
 }
