+++ web/src/pages/404.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import Layout from '../components/Layout'
-import { useSiteMetadata } from '../components/hooks/site-meta'
+import { useSiteMetadata } from '../components/Hooks/SiteMeta'
 
 const NotFoundPage = () => {
   return (
